added set_packing
